Extracted import function into a separate file
